refactor(frontend): extract useAppPresenter from App component
Moves state management, route coordination, and submit orchestration out of App.tsx into the useAppPresenter custom hook, leaving App.tsx as a pure stateless view layout.
